+++ src/components/dashboard/type.ts
@@ -1,0 +1,8 @@
+export type DashBaordUserProfileProps = {
+  userProfileData?: {
+    username: string;
+    nickname: string;
+    profile_image: string;
+    level: string;
+  };
+};
